+++ test/FirstDerivative/BasicArithmeticAll.C
@@ -7,27 +7,27 @@
 
 extern "C" int printf(const char* fmt, ...);
 
-float basic_1(int x) {
-  int y = 4;
-  int z = 3;
+float basic_1(float x) {
+  float y = 4;
+  float z = 3;
   return (y + x) / (x - z) * ((x * y * z) / 5); // == y * z * (x * x - 2 * x * z - y * z) / (5 * (x - z) * (x - z))
 }
-// CHECK: float basic_1_darg0(int x) {
-// CHECK-NEXT: int _d_x = 1;
-// CHECK-NEXT: int _d_y = 0;
-// CHECK-NEXT: int y = 4;
-// CHECK-NEXT: int _d_z = 0;
-// CHECK-NEXT: int z = 3;
-// CHECK-NEXT: int _t0 = (y + x);
-// CHECK-NEXT: int _t1 = (x - z);
-// CHECK-NEXT: int _t2 = x * y;
-// CHECK-NEXT: int _t3 = (_t2 * z);
-// CHECK-NEXT: int _t4 = _t0 / _t1;
-// CHECK-NEXT: int _t5 = (_t3 / 5);
+// CHECK: float basic_1_darg0(float x) {
+// CHECK-NEXT: float _d_x = 1;
+// CHECK-NEXT: float _d_y = 0;
+// CHECK-NEXT: float y = 4;
+// CHECK-NEXT: float _d_z = 0;
+// CHECK-NEXT: float z = 3;
+// CHECK-NEXT: float _t0 = (y + x);
+// CHECK-NEXT: float _t1 = (x - z);
+// CHECK-NEXT: float _t2 = x * y;
+// CHECK-NEXT: float _t3 = (_t2 * z);
+// CHECK-NEXT: float _t4 = _t0 / _t1;
+// CHECK-NEXT: float _t5 = (_t3 / 5);
 // CHECK-NEXT: return (((_d_y + _d_x) * _t1 - _t0 * (_d_x - _d_z)) / (_t1 * _t1)) * _t5 + _t4 * ((((_d_x * y + x * _d_y) * z + _t2 * _d_z) * 5 - _t3 * 0) / (5 * 5));
 // CHECK-NEXT: }
 
-float basic_1_darg0(int x);
+float basic_1_darg0(float x);
 
 double fn1(double i, double j) {
   double t = 1;
@@ -98,7 +98,7 @@ double fn3(double i, double j) {
 
 int main () {
   clad::differentiate(basic_1, 0);
-  printf("Result is = %f\n", basic_1_darg0(1)); // CHECK-EXEC: Result is = -6
+  printf("Result is = %.2f\n", basic_1_darg0(1)); // CHECK-EXEC: Result is = -10.20
   INIT(fn1, "i");
   INIT(fn2, "i");
   INIT(fn3, "i");
