+++ test/FirstDerivative/ClassMethodCall.C
@@ -11,60 +11,60 @@ public:
   virtual ~A() {}
   A() {}
 
-   __attribute__((always_inline)) int f(int x) {
+   __attribute__((always_inline)) float f(float x) {
     return x;
   }
 
-  //CHECK:{{[__attribute__((always_inline)) ]*}}int f_darg0(int x){{[ __attribute__((always_inline))]*}} {
-  //CHECK-NEXT:       int _d_x = 1;
+  //CHECK:{{[__attribute__((always_inline)) ]*}}float f_darg0(float x){{[ __attribute__((always_inline))]*}} {
+  //CHECK-NEXT:       float _d_x = 1;
   //CHECK-NEXT:       A _d_this_obj;
   //CHECK-NEXT:       A *_d_this = &_d_this_obj;
   //CHECK-NEXT:       return _d_x;
   //CHECK-NEXT:   }
 
 
-  int g_1(int x, int y) {
+  float g_1(float x, float y) {
     return x*x + y;
   }
 
-  //CHECK:   int g_1_darg0(int x, int y) {
-  //CHECK-NEXT:       int _d_x = 1;
-  //CHECK-NEXT:       int _d_y = 0;
+  //CHECK:   float g_1_darg0(float x, float y) {
+  //CHECK-NEXT:       float _d_x = 1;
+  //CHECK-NEXT:       float _d_y = 0;
   //CHECK-NEXT:       A _d_this_obj;
   //CHECK-NEXT:       A *_d_this = &_d_this_obj;
   //CHECK-NEXT:       return _d_x * x + x * _d_x + _d_y;
   //CHECK-NEXT:   }
 
 
-  //CHECK:   int g_1_darg1(int x, int y) {
-  //CHECK-NEXT:       int _d_x = 0;
-  //CHECK-NEXT:       int _d_y = 1;
+  //CHECK:   float g_1_darg1(float x, float y) {
+  //CHECK-NEXT:       float _d_x = 0;
+  //CHECK-NEXT:       float _d_y = 1;
   //CHECK-NEXT:       A _d_this_obj;
   //CHECK-NEXT:       A *_d_this = &_d_this_obj;
   //CHECK-NEXT:       return _d_x * x + x * _d_x + _d_y;
   //CHECK-NEXT:   }
 
-  int g_2(int x, int y) {
+  float g_2(float x, float y) {
     return x + y*y;
   }
 
-  //CHECK:   int g_2_darg0(int x, int y) {
-  //CHECK-NEXT:       int _d_x = 1;
-  //CHECK-NEXT:       int _d_y = 0;
+  //CHECK:   float g_2_darg0(float x, float y) {
+  //CHECK-NEXT:       float _d_x = 1;
+  //CHECK-NEXT:       float _d_y = 0;
   //CHECK-NEXT:       A _d_this_obj;
   //CHECK-NEXT:       A *_d_this = &_d_this_obj;
   //CHECK-NEXT:       return _d_x + _d_y * y + y * _d_y;
   //CHECK-NEXT:   }
 
-  //CHECK:   int g_2_darg1(int x, int y) {
-  //CHECK-NEXT:       int _d_x = 0;
-  //CHECK-NEXT:       int _d_y = 1;
+  //CHECK:   float g_2_darg1(float x, float y) {
+  //CHECK-NEXT:       float _d_x = 0;
+  //CHECK-NEXT:       float _d_y = 1;
   //CHECK-NEXT:       A _d_this_obj;
   //CHECK-NEXT:       A *_d_this = &_d_this_obj;
   //CHECK-NEXT:       return _d_x + _d_y * y + y * _d_y;
   //CHECK-NEXT:   }
 
-  int m(int x, int y) {
+  float m(float x, float y) {
     return f(x) + g_1(x, y);
   }
